Wire tx/receipts root
