+++ basic_bootloader/src/bootloader/block_flow/zk/post_tx_op/post_tx_op_proving_singleblock_batch.rs
@@ -59,7 +59,8 @@ where
     ) -> Result<Self::PostTxLoopOpResult, BootloaderSubsystemError> {
         let block_header = form_block_header(
             &system,
-            block_data.transaction_hashes_accumulator.finish().0,
+            block_data.transactions_root(),
+            block_data.receipts_root(),
             block_data.block_gas_used,
         )?;
         let block_hash = Bytes32::from(block_header.hash());
